+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_pos/AnnotatePositionCommand.java
@@ -8,8 +8,8 @@ import org.apache.commons.cli.ParseException;
 import de.charite.compbio.jannovar.JannovarException;
 import de.charite.compbio.jannovar.JannovarOptions;
 import de.charite.compbio.jannovar.annotation.AllAnnotationListTextGenerator;
-import de.charite.compbio.jannovar.annotation.AnnotationList;
-import de.charite.compbio.jannovar.annotation.AnnotationListTextGenerator;
+import de.charite.compbio.jannovar.annotation.VariantAnnotations;
+import de.charite.compbio.jannovar.annotation.VariantAnnotationsTextGenerator;
 import de.charite.compbio.jannovar.annotation.BestAnnotationListTextGenerator;
 import de.charite.compbio.jannovar.annotation.VariantAnnotator;
 import de.charite.compbio.jannovar.annotation.builders.AnnotationBuilderOptions;
@@ -58,7 +58,7 @@ public class AnnotatePositionCommand extends JannovarAnnotationCommand {
 			final GenomeVariant genomeChange = parseGenomeChange(chromosomalChange);
 
 			// Construct VariantAnnotator for building the variant annotations.
-			AnnotationList annoList = null;
+			VariantAnnotations annoList = null;
 			try {
 				annoList = annotator.buildAnnotationList(genomeChange);
 			} catch (Exception e) {
@@ -69,7 +69,7 @@ public class AnnotatePositionCommand extends JannovarAnnotationCommand {
 			// Obtain first or all functional annotation(s) and effect(s).
 			final String annotation;
 			final String effect;
-			AnnotationListTextGenerator textGenerator;
+			VariantAnnotationsTextGenerator textGenerator;
 			if (options.showAll)
 				textGenerator = new AllAnnotationListTextGenerator(annoList, 0, 1);
 			else